Restrict build to Linux platforms only
libstoragemgmt is a Red Hat funded project. While it is good to have
support for other architectures/platforms, and the same has been
expressed by upstream, for now, due to lack of time, I am marking it
to linux only.

Upstream's prime focus will be linux platforms. If, in future, I have
time, or else if someone else collaborates to co-maintian, we can
always revisit to enable these platforms.